Convert rickety bin scripts to legitimate CLI
Using Thor instead of stray Ruby code in scripts gives us more robust
option-parsing and documentation, and makes it easier to test this code. It will
also make easier follow-on work to compose a pipeline of extractors,
transformers, and loaders.

Also, switch from Coveralls to CodeClimate. Coveralls ties us to an outdated
version of Thor.