Require rdf library in entity resolver
Else, successful resolutions barf when encountering `RDF::URI.new(resp.body)` because of an uninitialized constant. I did not add a test for this behavior because RDF.rb is already required in the test suite and it did not seem worth trying to e.g. unrequire it to create a failure.