Redesign WoS extraction
Refs sul-dlss/rialto#271

This branch commits a substantial redesign of the WoS extractor and the accompanying WoS API client. With the prior design, the extractor was failing to extract more than a few dozen records at a time due to the API balking at the number of connections to its user query API. Instead, the client should hit the user query API once per researcher and use the query ID that is returned by that request to issue subsequent queries about said researcher. In order to accommodate this change, I found it helpful to isolate knowledge of the API to the service client, making the extractor significantly dumber and the client significantly smarter. The client now acts like an enumerable, responding to `#each`, and preventing knowledge of the WoS API (e.g., dealing with throttled connections, knowing which endpoints to hit for what purposes) from leaking into the extractor.